refactors a bunch of stuff out of Insights.tsx into a separate component
